detail_sub_page function and edit function done, the base set model is able to use easily.
