+++ lib/components/left_drawer/left_drawer.dart
@@ -3,7 +3,7 @@ import 'package:flutter_redux/flutter_redux.dart';
 import 'package:redux/redux.dart';
 import 'package:writing_writer2/components/toast_dialog.dart';
 import 'package:writing_writer2/server/file/IOBase.dart';
-import '../redux/app_state/state.dart';
+import '../../redux/app_state/state.dart';
 import 'book_listview.dart';
 
 class LeftDrawer extends StatefulWidget {
