+++ lib/components/right_drawer/right_drawer.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 import 'package:redux/redux.dart';
-import 'package:writing_writer2/components/set_listview.dart';
+import 'package:writing_writer2/components/right_drawer/set_listview.dart';
 import 'package:writing_writer2/components/toast_dialog.dart';
-import '../redux/app_state/state.dart';
+import '../../redux/app_state/state.dart';
 
 class RightDrawer extends StatefulWidget {
   const RightDrawer({Key? key,}) : super(key: key);
@@ -18,14 +18,14 @@ class _RightDrawerState extends State<RightDrawer> {
   Widget build(BuildContext context) {
     return Drawer(
       width: MediaQuery.of(context).size.width / 4.0,
-      child: StoreConnector<AppState, Map>(
+      child: StoreConnector<AppState, Map<String, dynamic>>(
         converter: (Store store) {
           return {
             'currentBookName': store.state.textModel.currentBook,
             'ioBase': store.state.ioBase,
           };
         },
-        builder: (BuildContext context, Map map) {
+        builder: (BuildContext context, Map<String, dynamic> map) {
           return Scaffold(
             appBar: AppBar(
               centerTitle: true,
