+++ lib/components/right_drawer/settings_listview.dart
@@ -1,23 +1,23 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 import 'package:redux/redux.dart';
-import '../redux/action/text_action.dart';
-import '../redux/app_state/state.dart';
+import '../../redux/action/set_action.dart';
+import '../../redux/app_state/state.dart';
 
-class ChapterListView extends StatefulWidget {
-  final String bookName;
+class SettingsListView extends StatefulWidget {
+  final String setName;
 
-  const ChapterListView({
+  const SettingsListView({
     Key? key,
-    required this.bookName,
+    required this.setName,
   }) : super(key: key);
 
   @override
-  State<ChapterListView> createState() => _ChapterListViewState();
+  State<SettingsListView> createState() => _SettingsListViewState();
 }
 
-class _ChapterListViewState extends State<ChapterListView> {
-  List<Widget> chapterListViewItems = [];
+class _SettingsListViewState extends State<SettingsListView> {
+  List<Widget> settingsListViewItems = [];
 
   @override
   void initState() {
@@ -25,21 +25,21 @@ class _ChapterListViewState extends State<ChapterListView> {
   }
 
   List<Widget> createChapterList(List<String> chapterList) {
-    chapterListViewItems.clear();
-    for (var chapterName in chapterList) {
-      chapterListViewItems.add(ChapterListViewItem(
-        bookName: widget.bookName,
-        chapterName: chapterName,
+    settingsListViewItems.clear();
+    for (var settingName in chapterList) {
+      settingsListViewItems.add(SettingsListViewItem(
+        setName: widget.setName,
+        settingName: settingName,
       ));
     }
-    return chapterListViewItems;
+    return settingsListViewItems;
   }
 
   @override
   Widget build(BuildContext context) {
     return StoreConnector<AppState, List<String>>(
       converter: (Store store) {
-        return store.state.ioBase.getAllChapters(widget.bookName);
+        return store.state.ioBase.getAllSettings(store.state.textModel.currentBook, widget.setName);
       },
       builder: (BuildContext context, List<String> chapterList) {
         return Column(
@@ -50,14 +50,14 @@ class _ChapterListViewState extends State<ChapterListView> {
   }
 }
 
-class ChapterListViewItem extends StatelessWidget {
-  final String bookName;
-  final String chapterName;
+class SettingsListViewItem extends StatelessWidget {
+  final String setName;
+  final String settingName;
 
-  const ChapterListViewItem({
+  const SettingsListViewItem({
     Key? key,
-    required this.bookName,
-    required this.chapterName,
+    required this.setName,
+    required this.settingName,
   }) : super(key: key);
 
   @override
@@ -67,11 +67,11 @@ class ChapterListViewItem extends StatelessWidget {
       converter: (Store store) {
         return () => {
           store.dispatch(
-            SetTextDataAction(currentBook: bookName, currentChapter: chapterName),
+            SetSetDataAction(currentSet: setName, currentSetting: settingName),
           ),
         };
       },
-      builder: (BuildContext context, VoidCallback clickChapter) {
+      builder: (BuildContext context, VoidCallback clickSetting) {
         return InkWell(
           child: Container(
             height: height / 18.0,
@@ -87,12 +87,12 @@ class ChapterListViewItem extends StatelessWidget {
             child: Row(
               mainAxisAlignment: MainAxisAlignment.center,
               children: [
-                Text(chapterName),
+                Text(settingName),
               ],
             ),
           ),
           onTap: () {
-            clickChapter();
+            clickSetting();
           },
         );
       },
